+++ Changelog.md
@@ -34,6 +34,7 @@
 - Fix: Taking an item out of a Display Case does not work when the offhand is not empty
 - Fix: Issues when unloading and saving chunks containing linkable block entities
 - Fix: Some scroll lists can be scrolled a tiny bit even though there is no scrollbar
+- Fix: Some mines explode when detecting a player in creative mode
 
 --------------------------Changelog for v1.9.4 of SecurityCraft--------------------------